requirements.txt
astroid==2.11.5
atomicwrites==1.4.0
attrs==21.4.0
click==8.1.3
colorama==0.4.4
dill==0.3.5.1
Flask==2.1.2
importlib-metadata==4.11.4
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pylint==2.13.9
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
tomli==2.0.1
typing_extensions==4.2.0
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
